project/requirements/common.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.1
Django==3.0.5
django-admin-sortable==2.2.3
django-countries==6.1.3
django-js-asset==1.2.2
django-model-utils==4.0.0
kombu==4.6.11
mysqlclient==1.4.6
Pillow==7.2.0
pytz==2020.1
redis==3.4.1
sqlparse==0.3.1
vine==1.3.0
django-summernote==0.8.11.6
django-modeladmin-reorder==0.3.1
django-middleware-global-request==0.2.0
django-solo==1.1.3
lxml==4.5.0
authorizenet==1.1.3
Cerberus==1.3.2
pycrypto==2.6.1
